Re-organise solution structure
