+++ HaveIBeenPwned/BreachCheckers/BaseChecker.cs
@@ -6,7 +6,7 @@ using System.Drawing;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 
-namespace HaveIBeenPwned
+namespace HaveIBeenPwned.BreachCheckers
 {
     public abstract class BaseChecker
     {
