+++ HaveIBeenPwned/BreachCheckers/Cloudbleed/CloudbleedEntry.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace HaveIBeenPwned
+namespace HaveIBeenPwned.BreachCheckers.Cloudbleed
 {
     public class CloudbleedEntry : IBreach
     {
