+++ HaveIBeenPwned/BreachedEntriesDialog.cs
@@ -1,4 +1,6 @@
-﻿using KeePass.Plugins;
+﻿using HaveIBeenPwned.BreachCheckers;
+using HaveIBeenPwned.Extensions;
+using KeePass.Plugins;
 using KeePassLib;
 using System;
 using System.Collections.Generic;
