+++ HaveIBeenPwned/CheckerPrompt.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using HaveIBeenPwned.Extensions;
+using System;
 using System.Drawing;
 using System.Linq;
 using System.Windows.Forms;
