+++ HaveIBeenPwned/HaveIBeenPwned.csproj
@@ -51,14 +51,14 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="BaseChecker.cs" />
+    <Compile Include="BreachCheckers\BaseChecker.cs" />
     <Compile Include="BreachedEntriesDialog.cs">
       <SubType>Form</SubType>
     </Compile>
     <Compile Include="BreachedEntriesDialog.Designer.cs">
       <DependentUpon>BreachedEntriesDialog.cs</DependentUpon>
     </Compile>
-    <Compile Include="BreachedEntry.cs" />
+    <Compile Include="BreachCheckers\BreachedEntry.cs" />
     <Compile Include="BreachEnum.cs" />
     <Compile Include="CheckerPrompt.cs">
       <SubType>Form</SubType>
@@ -66,19 +66,19 @@
     <Compile Include="CheckerPrompt.Designer.cs">
       <DependentUpon>CheckerPrompt.cs</DependentUpon>
     </Compile>
-    <Compile Include="CloudbleedChecker.cs" />
-    <Compile Include="CloudbleedEntry.cs" />
+    <Compile Include="BreachCheckers\Cloudbleed\CloudbleedChecker.cs" />
+    <Compile Include="BreachCheckers\Cloudbleed\CloudbleedEntry.cs" />
     <Compile Include="DisplayAttribute.cs" />
-    <Compile Include="EnumExtensions.cs" />
-    <Compile Include="HaveIBeenPwnedChecker.cs" />
-    <Compile Include="HaveIBeenPwnedEntry.cs" />
+    <Compile Include="Extensions\EnumExtensions.cs" />
+    <Compile Include="BreachCheckers\HaveIBeenPwned\HaveIBeenPwnedChecker.cs" />
+    <Compile Include="BreachCheckers\HaveIBeenPwned\HaveIBeenPwnedEntry.cs" />
     <Compile Include="HaveIBeenPwnedExt.cs" />
-    <Compile Include="IBreach.cs" />
+    <Compile Include="BreachCheckers\IBreach.cs" />
     <Compile Include="ProgressHelper.cs" />
     <Compile Include="ProgressItem.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="ProtectedStringExtensions.cs" />
-    <Compile Include="PwEntryExtensions.cs" />
+    <Compile Include="Extensions\ProtectedStringExtensions.cs" />
+    <Compile Include="Extensions\PwEntryExtensions.cs" />
     <Compile Include="Resources.Designer.cs">
       <AutoGen>True</AutoGen>
       <DesignTime>True</DesignTime>
